# backend/openapi_dump.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "HOSPITALITY AI OS",
                 "description":  "Enterprise AI Operating System for Premium Hospitality",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/v1/integrations/webhook/toast":  {
                                                             "post":  {
                                                                          "tags":  [
                                                                                       "integrations"
                                                                                   ],
                                                                          "summary":  "Receive Toast Order",
                                                                          "description":  "Webhook endpoint to receive orders from Toast POS.\nNormalization happens here.",
                                                                          "operationId":  "receive_toast_order_api_v1_integrations_webhook_toast_post",
                                                                          "requestBody":  {
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/Order"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          },
                                                                                              "required":  true
                                                                                          },
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "Successful Response",
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/Order"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    }
                                                                                                    },
                                                                                            "422":  {
                                                                                                        "description":  "Validation Error",
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    }
                                                                                                    }
                                                                                        }
                                                                      }
                                                         },
...
                  "/api/v1/integrations/guests/{guest_id}":  {
                                                                 "get":  {
                                                                             "tags":  [
                                                                                          "integrations"
                                                                                      ],
                                                                             "summary":  "Get Guest Profile",
                                                                             "description":  "Retrieve unified guest profile.",
                                                                             "operationId":  "get_guest_profile_api_v1_integrations_guests__guest_id__get",
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "guest_id",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string",
                                                                                                                   "title":  "Guest Id"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "Successful Response",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/GuestProfile"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       },
                                                                                               "422":  {
                                                                                                           "description":  "Validation Error",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "Order":  {
                                                     "properties":  {
                                                                        "id":  {
                                                                                   "type":  "string",
                                                                                   "title":  "Id"
                                                                               },
                                                                        "source":  {
                                                                                       "type":  "string",
                                                                                       "title":  "Source",
                                                                                       "default":  "toast"
                                                                                   },
                                                                        "table_number":  {
                                                                                             "anyOf":  [
                                                                                                           {
                                                                                                               "type":  "integer"
                                                                                                           },
                                                                                                           {
                                                                                                               "type":  "null"
                                                                                                           }
                                                                                                       ],
                                                                                             "title":  "Table Number"
                                                                                         },
                                                                        "guest_count":  {
                                                                                            "type":  "integer",
                                                                                            "title":  "Guest Count",
                                                                                            "default":  1
                                                                                        },
                                                                        "items":  {
                                                                                      "items":  {
                                                                                                    "$ref":  "#/components/schemas/OrderItem"
                                                                                                },
                                                                                      "type":  "array",
                                                                                      "title":  "Items"
                                                                                  },
                                                                        "total_amount":  {
                                                                                             "type":  "number",
                                                                                             "title":  "Total Amount"
                                                                                         },
                                                                        "status":  {
                                                                                       "type":  "string",
                                                                                       "title":  "Status",
                                                                                       "default":  "open"
                                                                                   },
                                                                        "created_at":  {
                                                                                           "type":  "string",
                                                                                           "format":  "date-time",
                                                                                           "title":  "Created At"
                                                                                       }
                                                                    },
                                                     "type":  "object",
                                                     "required":  [
                                                                      "id",
                                                                      "items",
                                                                      "total_amount"
                                                                  ],
                                                     "title":  "Order"
                                                 },
...
                                   }
                   }
}